Move SimpleDatabaseConnection, update imports
