+++ src/main/java/database/SimpleDatabaseConnection.java
@@ -1,6 +1,5 @@
-package tools.mapletools;
+package database;
 
-import database.DatabaseConnection;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.config.Configurator;
