+++ src/main/java/tools/mapletools/CashDropFetcher.java
@@ -1,5 +1,6 @@
 package tools.mapletools;
 
+import database.SimpleDatabaseConnection;
 import provider.wz.WZFiles;
 import tools.Pair;
