+++ src/main/java/tools/mapletools/GachaponItemIdRetriever.java
@@ -1,5 +1,7 @@
 package tools.mapletools;
 
+import database.SimpleDatabaseConnection;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.PrintWriter;
