+++ src/main/java/tools/mapletools/IdRetriever.java
@@ -1,5 +1,7 @@
 package tools.mapletools;
 
+import database.SimpleDatabaseConnection;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
