+++ src/main/java/tools/mapletools/MobBookIndexer.java
@@ -1,5 +1,6 @@
 package tools.mapletools;
 
+import database.SimpleDatabaseConnection;
 import provider.wz.WZFiles;
 
 import java.io.BufferedReader;
