+++ src/main/java/tools/mapletools/SkillbookChanceFetcher.java
@@ -1,5 +1,6 @@
 package tools.mapletools;
 
+import database.SimpleDatabaseConnection;
 import server.life.MonsterStats;
 import tools.Pair;
